Review - Split the tests & rely on setup for method configuration
